因为数据库int存不了16位需要bigint后端对应long，已经自动成成和重新导包了

+++ src/main/java/edu/neu/hospital/bean/baseBean/UserView.java
@@ -76,7 +76,7 @@ public class UserView implements Serializable {
 
     private String photoLocation;
 
-    private Integer contact;
+    private Long contact;
 
     /**
      * 登录名
@@ -242,11 +242,11 @@ public class UserView implements Serializable {
         this.photoLocation = photoLocation;
     }
 
-    public Integer getContact() {
+    public Long getContact() {
         return contact;
     }
 
-    public void setContact(Integer contact) {
+    public void setContact(Long contact) {
         this.contact = contact;
     }
